add_context_nav → Views::FullPageBase::ContextNav
The last batch of the `content_for` helper sweep. 128 action-template
call sites switch from helper proxy to direct method dispatch.
`add_context_nav` + the private `normalize_context_nav_links` move
into a new `Views::FullPageBase::ContextNav` mixin alongside the
other per-concern modules. The setter's body is otherwise unchanged
beyond the obligatory `capture { render(...) }` wrap for Phlex 2's
emit-to-buffer `render` semantics.

The helper's `context_nav_links` / `context_nav_link` /
`merge_context_nav_link_args` / `crud_button_or_link` family had a
single non-helper caller: `Views::Controllers::Users::Show::Profile#
render_action_links`. Restructured to include the existing
`Views::Layouts::ContextNav::LinkRendering` module (the Phlex-side
dispatch the two `*::ContextNav` views already share) and call
`render_crud_button_or_link` directly inside an `<li>`. Net: ~95
lines of duplicated logic in the helper go.

`app/helpers/header/context_nav_helper.rb` deleted.
`register_value_helper :add_context_nav` + `:context_nav_links` lines
in `Views::Base` removed.

Verified locally with 204-test strategic sweep across obs / names /
herbaria / users index + show pages — `users` covers the
`profile.rb` rework.
